// src/app/shared/config/app-url.enum.ts
export enum AppUrl {
   login = 'user/v1/user/login',
   userList = 'user/v1/user/list',
   userImage = 'user/v1/user/pic',
   loggedUser = 'user/v1/user',
   filters = 'user/v1/user/locationDepartmentlist',
   upload = 'user/v1/user/file',
   emailList = 'user/v1/user/email/list',
   saveDog = 'dog/v1/add',
   updateDog = 'dog/v1/update',
   deleteDog = 'dog/v1/delete'

}